[sim_api] Add cast from char* to unsigned long in SimNamedMarker to avoid compiler warnings

+++ include/sim_api.h
@@ -85,7 +85,7 @@
 #define SimGetFreqMHz(proc)       SimMagic1(SIM_CMD_MHZ_GET, proc)
 #define SimGetOwnFreqMHz()        SimGetFreqMHz(UINT64_MAX)
 #define SimMarker(arg0, arg1)     SimMagic2(SIM_CMD_MARKER, arg0, arg1)
-#define SimNamedMarker(arg0, str) SimMagic2(SIM_CMD_NAMED_MARKER, arg0, str)
+#define SimNamedMarker(arg0, str) SimMagic2(SIM_CMD_NAMED_MARKER, arg0, (unsigned long)(str))
 #define SimUser(cmd, arg)         SimMagic2(SIM_CMD_USER, cmd, arg)
 #define SimSetInstrumentMode(opt) SimMagic1(SIM_CMD_INSTRUMENT_MODE, opt)
 #define SimInSimulator()          (SimMagic0(SIM_CMD_IN_SIMULATOR)!=SIM_CMD_IN_SIMULATOR)
